+++ Week2/src/context/Theme/ThemeProvider.tsx
@@ -1,0 +1,26 @@
+import { PropsWithChildren, useEffect, useState } from "react";
+import { THEME, TTheme } from "./ThemeType";
+import { ThemeContext } from "./ThemeContext";
+
+export const ThemeProvider = ({ children }: PropsWithChildren) => {
+  const [theme, setTheme] = useState<TTheme>(THEME.LIGHT);
+
+  const toggleTheme = () => {
+    setTheme((prev) => (prev === THEME.LIGHT ? THEME.DARK : THEME.LIGHT));
+  };
+
+  useEffect(() => {
+    const root = document.documentElement; // <html>
+    if (theme === THEME.DARK) {
+      root.classList.add("dark");
+    } else {
+      root.classList.remove("dark");
+    }
+  }, [theme]);
+
+  return (
+    <ThemeContext.Provider value={{ theme, toggleTheme }}>
+      {children}
+    </ThemeContext.Provider>
+  );
+};
