+++ Week2/src/context/Theme/ThemeType.tsx
@@ -1,0 +1,6 @@
+export enum THEME {
+  LIGHT = "LIGHT",
+  DARK = "DARK",
+}
+
+export type TTheme = THEME.LIGHT | THEME.DARK;
