--- Week2/src/context/ThemeProvider.tsx
@@ -1,57 +1,0 @@
-/* eslint-disable react-refresh/only-export-components */
-import {
-  createContext,
-  PropsWithChildren,
-  useContext,
-  useEffect,
-  useState,
-} from "react";
-
-export enum THEME {
-  LIGHT = "LIGHT",
-  DARK = "DARK",
-}
-
-type TTheme = THEME.DARK | THEME.LIGHT;
-
-interface IThemeContext {
-  theme: TTheme;
-  toggleTheme: () => void;
-}
-
-export const ThemeContext = createContext<IThemeContext | undefined>(undefined);
-
-export const ThemeProvider = ({ children }: PropsWithChildren) => {
-  const [theme, setTheme] = useState<TTheme>(THEME.LIGHT);
-
-  const toggleTheme = () => {
-    setTheme((prev) => (prev === THEME.LIGHT ? THEME.DARK : THEME.LIGHT));
-  };
-
-  useEffect(() => {
-    const root = document.documentElement; // <html>
-    if (theme === THEME.DARK) {
-      root.classList.add("dark");
-    } else {
-      root.classList.remove("dark");
-    }
-  }, [theme]);
-
-  return (
-    <ThemeContext.Provider value={{ theme, toggleTheme }}>
-      {children}
-    </ThemeContext.Provider>
-  );
-};
-
-export const useTheme = () => {
-  const context = useContext(ThemeContext);
-  // 컨텍스트가 없는 경우 처리
-  if (!context) {
-    throw new Error(
-      "useTheme 반드시 ThemeProvider 내부에서 사용되어야 합니다."
-    );
-  }
-  // 컨텍스트가 있는 경우
-  return context;
-};
